# disz-backend/swagger/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost","basePath":"/","tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"book-controller","description":"Book Controller"},{"name":"borrow-controller","description":"Borrow Controller"},{"name":"person-controller","description":"Person Controller"}],"paths":{"/books":{"get":{"tags":["book-controller"],"summary":"getBooks","operationId":"getBooksUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BookDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["book-controller"],"summary":"addBook","operationId":"addBookUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"book","description":"book","required":true,"schema":{"$ref":"#/definitions/BookDto"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/books/borrowed":{"get":{"tags":["book-controller"],"summary":"findByPersonIdIsNotNull","operationId":"findByPersonIdIsNotNullUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BookDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/books/search":{"get":{"tags":["book-controller"],"summary":"findBooks","operationId":"findBooksUsingGET","produces":["*/*"],"parameters":[{"name":"query","in":"query","description":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BookDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/books/{id}":{"get":{"tags":["book-controller"],"summary":"getBook","operationId":"getBookUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BookDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["book-controller"],"summary":"updateBook","operationId":"updateBookUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"author","in":"query","description":"author","required":false,"type":"string"},{"in":"body","name":"book","description":"book","required":true,"schema":{"$ref":"#/definitions/BookDto"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"title","in":"query","description":"title","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BookDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["book-controller"],"summary":"deleteBook","operationId":"deleteBookUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/borrows":{"get":{"tags":["borrow-controller"],"summary":"findAll","operationId":"findAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["borrow-controller"],"summary":"addBorrow","operationId":"addBorrowUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"borrowDto","description":"borrowDto","required":true,"schema":{"$ref":"#/definitions/BorrowDto"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/books/{bookId}":{"get":{"tags":["borrow-controller"],"summary":"findBorrowByBookId","operationId":"findBorrowByBookIdUsingGET","produces":["*/*"],"parameters":[{"name":"bookId","in":"path","description":"bookId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/borrowed":{"get":{"tags":["borrow-controller"],"summary":"findBorrowByStartTimeIsNotNullAndEndTimeIsNull","operationId":"findBorrowByStartTimeIsNotNullAndEndTimeIsNullUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/borrowed/count":{"get":{"tags":["borrow-controller"],"summary":"nowBorrowedBooks","operationId":"nowBorrowedBooksUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"integer","format":"int32"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/person/{personId}":{"get":{"tags":["borrow-controller"],"summary":"findBorrowByPersonId","operationId":"findBorrowByPersonIdUsingGET","produces":["*/*"],"parameters":[{"name":"personId","in":"path","description":"personId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/returned":{"get":{"tags":["borrow-controller"],"summary":"findBorrowByStartTimeAndEndTimeIsNotNull","operationId":"findBorrowByStartTimeAndEndTimeIsNotNullUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/returned/count":{"get":{"tags":["borrow-controller"],"summary":"returnesBorrowes","operationId":"returnesBorrowesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"integer","format":"int32"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/search":{"get":{"tags":["borrow-controller"],"summary":"findByAuthorOrTitleContaining","operationId":"findByAuthorOrTitleContainingUsingGET","produces":["*/*"],"parameters":[{"name":"search","in":"query","description":"search","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BorrowDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/borrows/{id}/return":{"put":{"tags":["borrow-controller"],"summary":"returnBook","operationId":"returnBookUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"borrowDto","description":"borrowDto","required":true,"schema":{"$ref":"#/definitions/BorrowDto"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/person":{"get":{"tags":["person-controller"],"summary":"getPersons","operationId":"getPersonsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PersonDtoRes"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["person-controller"],"summary":"addPerson","operationId":"addPersonUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"personDto","description":"personDto","required":true,"schema":{"$ref":"#/definitions/PersonDtoReq"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["person-controller"],"summary":"updatePerson","operationId":"updatePersonUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"newEmail","in":"query","description":"newEmail","required":false,"type":"string"},{"name":"newFirstName","in":"query","description":"newFirstName","required":false,"type":"string"},{"name":"newLastName","in":"query","description":"newLastName","required":false,"type":"string"},{"in":"body","name":"personDto","description":"personDto","required":true,"schema":{"$ref":"#/definitions/PersonDtoReq"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/search":{"get":{"tags":["person-controller"],"summary":"findPerson","operationId":"findPersonUsingGET","produces":["*/*"],"parameters":[{"name":"query","in":"query","description":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PersonDtoRes"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/{id":{"delete":{"tags":["person-controller"],"summary":"deletePerson","operationId":"deletePersonUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/person/{id}":{"get":{"tags":["person-controller"],"summary":"getPerson","operationId":"getPersonUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/PersonDtoRes"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/{password}":{"post":{"tags":["person-controller"],"summary":"pswChange","operationId":"pswChangeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"password","in":"path","description":"password","required":true,"type":"string"},{"in":"body","name":"personDto","description":"personDto","required":true,"schema":{"$ref":"#/definitions/PersonDtoReq"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/persons":{"get":{"tags":["person-controller"],"summary":"personsBooks","operationId":"personsBooksUsingGET","produces":["*/*"],"parameters":[{"name":"personDto","in":"path","description":"personDto","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Borrow"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Book":{"type":"object","properties":{"author":{"type":"string"},"borrows":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"id":{"type":"integer","format":"int64"},"title":{"type":"string"}},"title":"Book"},"BookDto":{"type":"object","properties":{"author":{"type":"string"},"bookId":{"type":"integer","format":"int64"},"borrows":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"title":{"type":"string"}},"title":"BookDto"},"Borrow":{"type":"object","properties":{"book":{"$ref":"#/definitions/Book"},"endTime":{"type":"string","format":"date"},"id":{"type":"integer","format":"int64"},"person":{"$ref":"#/definitions/Person"},"startTime":{"type":"string","format":"date"}},"title":"Borrow"},"BorrowDto":{"type":"object","properties":{"book":{"$ref":"#/definitions/Book"},"endTime":{"type":"string","format":"date"},"person":{"$ref":"#/definitions/Person"},"startTime":{"type":"string","format":"date"}},"title":"BorrowDto"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Person":{"type":"object","properties":{"admin":{"type":"boolean"},"borrows":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int64"},"lastName":{"type":"string"},"password":{"type":"string"}},"title":"Person"},"PersonDtoReq":{"type":"object","properties":{"admin":{"type":"boolean"},"borrows":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"personID":{"type":"integer","format":"int64"}},"title":"PersonDtoReq"},"PersonDtoRes":{"type":"object","properties":{"admin":{"type":"boolean"},"bookDtos":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"borrows":{"type":"array","items":{"$ref":"#/definitions/Borrow"}},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"personID":{"type":"integer","format":"int64"},"personId":{"type":"integer","format":"int64"}},"title":"PersonDtoRes"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}