# disz-frontend/src/api/src/model/BookDto.js
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';
import Borrow from './Borrow';

/**
 * The BookDto model module.
 * @module model/BookDto
 * @version 1.0
 */
class BookDto {
    /**
     * Constructs a new <code>BookDto</code>.
     * @alias module:model/BookDto
     */
    constructor() { 
        
        BookDto.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>BookDto</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/BookDto} obj Optional instance to populate.
     * @return {module:model/BookDto} The populated <code>BookDto</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new BookDto();

            if (data.hasOwnProperty('author')) {
                obj['author'] = ApiClient.convertToType(data['author'], 'String');
            }
            if (data.hasOwnProperty('bookId')) {
                obj['bookId'] = ApiClient.convertToType(data['bookId'], 'Number');
            }
            if (data.hasOwnProperty('borrows')) {
                obj['borrows'] = ApiClient.convertToType(data['borrows'], [Borrow]);
            }
            if (data.hasOwnProperty('title')) {
                obj['title'] = ApiClient.convertToType(data['title'], 'String');
            }
        }
        return obj;
    }


}

/**
 * @member {String} author
 */
BookDto.prototype['author'] = undefined;

/**
 * @member {Number} bookId
 */
BookDto.prototype['bookId'] = undefined;

/**
 * @member {Array.<module:model/Borrow>} borrows
 */
BookDto.prototype['borrows'] = undefined;

/**
 * @member {String} title
 */
BookDto.prototype['title'] = undefined;






export default BookDto;