# disz-frontend/src/api/src/model/PersonDtoReq.js
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';
import Borrow from './Borrow';

/**
 * The PersonDtoReq model module.
 * @module model/PersonDtoReq
 * @version 1.0
 */
class PersonDtoReq {
    /**
     * Constructs a new <code>PersonDtoReq</code>.
     * @alias module:model/PersonDtoReq
     */
    constructor() { 
        
        PersonDtoReq.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>PersonDtoReq</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/PersonDtoReq} obj Optional instance to populate.
     * @return {module:model/PersonDtoReq} The populated <code>PersonDtoReq</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new PersonDtoReq();

            if (data.hasOwnProperty('admin')) {
                obj['admin'] = ApiClient.convertToType(data['admin'], 'Boolean');
            }
            if (data.hasOwnProperty('borrows')) {
                obj['borrows'] = ApiClient.convertToType(data['borrows'], [Borrow]);
            }
            if (data.hasOwnProperty('email')) {
                obj['email'] = ApiClient.convertToType(data['email'], 'String');
            }
            if (data.hasOwnProperty('firstName')) {
                obj['firstName'] = ApiClient.convertToType(data['firstName'], 'String');
            }
            if (data.hasOwnProperty('lastName')) {
                obj['lastName'] = ApiClient.convertToType(data['lastName'], 'String');
            }
            if (data.hasOwnProperty('password')) {
                obj['password'] = ApiClient.convertToType(data['password'], 'String');
            }
            if (data.hasOwnProperty('personID')) {
                obj['personID'] = ApiClient.convertToType(data['personID'], 'Number');
            }
        }
        return obj;
    }


}

/**
 * @member {Boolean} admin
 */
PersonDtoReq.prototype['admin'] = undefined;

/**
 * @member {Array.<module:model/Borrow>} borrows
 */
PersonDtoReq.prototype['borrows'] = undefined;

/**
 * @member {String} email
 */
PersonDtoReq.prototype['email'] = undefined;

/**
 * @member {String} firstName
 */
PersonDtoReq.prototype['firstName'] = undefined;

/**
 * @member {String} lastName
 */
PersonDtoReq.prototype['lastName'] = undefined;

/**
 * @member {String} password
 */
PersonDtoReq.prototype['password'] = undefined;

/**
 * @member {Number} personID
 */
PersonDtoReq.prototype['personID'] = undefined;






export default PersonDtoReq;