# disz-frontend/src/api/src/model/PersonDtoRes.js
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';
import Borrow from './Borrow';

/**
 * The PersonDtoRes model module.
 * @module model/PersonDtoRes
 * @version 1.0
 */
class PersonDtoRes {
    /**
     * Constructs a new <code>PersonDtoRes</code>.
     * @alias module:model/PersonDtoRes
     */
    constructor() { 
        
        PersonDtoRes.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>PersonDtoRes</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/PersonDtoRes} obj Optional instance to populate.
     * @return {module:model/PersonDtoRes} The populated <code>PersonDtoRes</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new PersonDtoRes();

            if (data.hasOwnProperty('admin')) {
                obj['admin'] = ApiClient.convertToType(data['admin'], 'Boolean');
            }
            if (data.hasOwnProperty('bookDtos')) {
                obj['bookDtos'] = ApiClient.convertToType(data['bookDtos'], [Borrow]);
            }
            if (data.hasOwnProperty('borrows')) {
                obj['borrows'] = ApiClient.convertToType(data['borrows'], [Borrow]);
            }
            if (data.hasOwnProperty('email')) {
                obj['email'] = ApiClient.convertToType(data['email'], 'String');
            }
            if (data.hasOwnProperty('firstName')) {
                obj['firstName'] = ApiClient.convertToType(data['firstName'], 'String');
            }
            if (data.hasOwnProperty('lastName')) {
                obj['lastName'] = ApiClient.convertToType(data['lastName'], 'String');
            }
            if (data.hasOwnProperty('password')) {
                obj['password'] = ApiClient.convertToType(data['password'], 'String');
            }
            if (data.hasOwnProperty('personID')) {
                obj['personID'] = ApiClient.convertToType(data['personID'], 'Number');
            }
            if (data.hasOwnProperty('personId')) {
                obj['personId'] = ApiClient.convertToType(data['personId'], 'Number');
            }
        }
        return obj;
    }


}

/**
 * @member {Boolean} admin
 */
PersonDtoRes.prototype['admin'] = undefined;

/**
 * @member {Array.<module:model/Borrow>} bookDtos
 */
PersonDtoRes.prototype['bookDtos'] = undefined;

/**
 * @member {Array.<module:model/Borrow>} borrows
 */
PersonDtoRes.prototype['borrows'] = undefined;

/**
 * @member {String} email
 */
PersonDtoRes.prototype['email'] = undefined;

/**
 * @member {String} firstName
 */
PersonDtoRes.prototype['firstName'] = undefined;

/**
 * @member {String} lastName
 */
PersonDtoRes.prototype['lastName'] = undefined;

/**
 * @member {String} password
 */
PersonDtoRes.prototype['password'] = undefined;

/**
 * @member {Number} personID
 */
PersonDtoRes.prototype['personID'] = undefined;

/**
 * @member {Number} personId
 */
PersonDtoRes.prototype['personId'] = undefined;






export default PersonDtoRes;